adding models files, index.js, continueStory.js, newStory.js, user.js
